Avoid creating buffer when locating bufnr
While debugging Java with nvim-jdtls, step debugging into source
associated with a dependency would result in an error locating the
treesitter parser.

The code at the top of `set_virtual_text` tries to locate the buffer
associated with the source of the dap stack frame by transforming
`source.path` to a URI with `uri_from_fname` before calling
`uri_to_bufnr` to locate the buffer. This works well for the common case
of file names, but fails for cases where `source.path` is already a URI.

When an existing URI is provided to `uri_from_fname`, it undergoes
another set of encoding and get prefixed with at `file://` scheme. This
new URI is not associated with any existing buffer, so `uri_to_bufnr`
creates a new buffer. Since the new buffer has no associated filetype,
the code attempts to identify the filetype based on the file extension
from the transformed URI. Since the filetype associated with the
extension is not correct, treesitter is unable to create a parser and
the function fails.

This commit changes the processing to attempt to locate the buffer by
calling `vim.fn.bufnr` before falling back to the `uri_from_fname` and
`uri_to_bufnr` combination. In the case of nvim-jdtls, `vim.fn.bufnr`
returns the correct buffer with its configured options and state.